requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0

# OpenAI and LLM related
openai>=1.0.0
mem0ai>=0.1.0

# Vector search
faiss-cpu>=1.7.4
# If you need GPU support, replace faiss-cpu with: faiss-gpu>=1.7.4

# NLP and evaluation metrics
nltk>=3.8.0
bert-score>=0.3.13
rouge-score>=0.1.2
sentence-transformers>=2.2.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
jinja2>=3.1.0

# Optional: If you want to use the notebooks (not required for src files)
# jupyter>=1.0.0
# ipykernel>=6.25.0